добавил поддержку паттерна Наблюдатель для обновления состояния игры, интегрировал таймер анимации и улучшил управление игровым процессом
