+++ task_2_3_1/src/main/java/vylegzhanin/snake/model/GameObserver.java
@@ -1,0 +1,12 @@
+package vylegzhanin.snake.model;
+
+/**
+ * Интерфейс для реализации паттерна Наблюдатель (Observer).
+ * Позволяет View подписываться на изменения в Модели (Game).
+ */
+public interface GameObserver {
+    /**
+     * Вызывается, когда состояние игры обновилось (тик таймера или счет).
+     */
+    void onGameStateChanged();
+}
